Feat: removed stale code
