--- packages/core/event/event.go
@@ -1,19 +1,0 @@
-package event
-
-// import (
-// 	"github.com/iotaledger/hive.go/core/generics/options"
-// )
-//
-// type Event struct {
-// 	*base
-// }
-//
-// func (e *Event) Hook(callback func(), opts ...options.Option[Hook]) (hook *Hook) {
-// 	return e.hook(callback, opts...)
-// }
-//
-// func (e *Event) Trigger() {
-// 	e.trigger(func(hook *Hook) {
-// 		hook.callback.(func())()
-// 	})
-// }
