+++ packages/core/event/with2params.go
@@ -1,44 +1,27 @@
 package event
 
-//
-// import (
-// 	"github.com/iotaledger/hive.go/core/generics/options"
-// )
-//
-// type With2Params[A, B any] struct {
-// 	*base
-// }
-//
-// func New2[A, B any]() *With2Params[A, B] {
-// 	return &With2Params[A, B]{
-// 		base: newBase(),
-// 	}
-// }
-//
-// func (w *With2Params[A, B]) Hook(callback func(a A, b B), opts ...options.Option[Hook]) (hook *Hook) {
-// 	return w.base.hook(callback, opts...)
-// }
-//
-// func (w *With2Params[A, B]) Trigger(a A, b B) {
-// 	if w.triggerCount.Inc() >= w.maxTriggerCount && w.maxTriggerCount > 0 {
-// 		return
-// 	}
-//
-// 	w.hooks.ForEach(func(_ uint64, hook *Hook) bool {
-// 		if hook.triggerCount.Inc() >= hook.maxTriggerCount && hook.maxTriggerCount > 0 {
-// 			return true
-// 		}
-//
-// 		if hook.workerPool == nil {
-// 			hook.callback.(func(A, B))(a, b)
-//
-// 			return true
-// 		}
-//
-// 		hook.workerPool.Submit(func() {
-// 			hook.callback.(func(A, B))(a, b)
-// 		})
-//
-// 		return true
-// 	})
-// }
+type With2Params[Param1, Param2 any] struct {
+	*base[func(Param1, Param2)]
+}
+
+func New2[Param1, Param2 any]() *With2Params[Param1, Param2] {
+	return &With2Params[Param1, Param2]{
+		base: newBase[func(Param1, Param2)](),
+	}
+}
+
+func (w *With2Params[Param1, Param2]) Trigger(param1 Param1, param2 Param2) {
+	if w.triggerCount.Add(1) < w.maxTriggerCount || w.maxTriggerCount == 0 {
+		w.hooks.ForEach(func(_ uint64, hook *Hook[func(Param1, Param2)]) bool {
+			if hook.shouldTrigger() {
+				if workerPool := w.targetWorkerPool(hook); workerPool == nil {
+					hook.callback(param1, param2)
+				} else {
+					workerPool.Submit(func() { hook.callback(param1, param2) })
+				}
+			}
+
+			return true
+		})
+	}
+}
